Fix a warning reported by static analyser tool: Assignment of a signed value which has type 'time_t' to a variable of a bigger integer type 'guint64'

+++ gio/gdbusauthmechanismsha1.c
@@ -658,7 +658,7 @@ keyring_generate_entry (const gchar  *cookie_context,
     }
 
   new_contents = g_string_new (NULL);
-  now = time (NULL);
+  now = (guint64) time (NULL);
   changed_file = FALSE;
 
   max_line_id = 0;
